pyproject.toml
[project]
name = "agentic_ai_mcp_demo"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.35.0",
    "boto3>=1.35.49",
    "langchain>=0.3.9",
    "langchain-chroma>=0.1.4",
    "langgraph>=0.2.53",
    "mcp>=1.0.0",
    "notebook>=7.2.2",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "streamlit>=1.39.0",
    "langchain-openai>=0.2.10",
    "langchain-community>=0.3.9",
    "langchain-anthropic>=0.2.4",
    "pyvis>=0.3.2",
    "langchain-mcp",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "icecream>=2.1.3"
]

[tool.uv.sources]
langchain-mcp = { path = "../langchain_mcp", editable = true }


[tool.ruff.lint]
select = [
    # pydocstyle
    "D",
    # Google style
    "G",
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # flake8-annotations
    "ANN",
]

ignore = [
    # flask8-annotations for self
    "ANN101",
    "ANN102",
    "ANN204",
    "G004",
    "D100",
    "D107"
]

exclude = [
    "*.ipynb",
    "__init__.py"
]


[tool.ruff.lint.pydocstyle]
convention = "google"
